Move functions from IrcBridge to ClientPool
